requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
xmlrpc==1.0.1
schedule==1.2.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9